add prerequisite for pro deployments
```
    "policies": {
        "allow_explicit_policy_id": true
    },
```

Is required to be set in `tyk.conf` for SecurityPolicy resources to properly
work.

This is because the operator uses `id` to refer to security policy id's while
the gateways will by default use `_id` to refer to policy id's and when `policies.allow_explicit_policy_id` is true the gateway will use `id` when
it is provided to identify policies.