docs(installation): update, and clarify some points
To be more clear and explicit, refer to the gateway as "Tyk Gateway" and the operator as "Tyk
Operator" where some places were just referring to something as "Tyk". Also straightened out the
TOCs in the base README and the installation doc to auto-generate cleanly against the
existing/updated heading lines.