avoid updating access_rights_array on SecurityPolicySpec (#303)
* apply crds on scrap

* avoid updating access_rights_array on SecurityPolicySpec

any updates on this object confused k8s diff even when the object didn't
change.

In this commit, we only update the copy that is sent to the dashboard for
Update and Create calls. The dashboard will receive the objects with
`api_name` and `api_id` properly set, however these will not be stored in
k8s since we already refer to the referenced api by name and namespace
and they serve no other purpose.

This means kubectl apply will return unchanged when the resource hasn't changed.