Improved the documentation regarding the version of cert-manager.
Description: with the 1.22 kubernetes update, MutatingWebhookConfiguration API has been removed, this led to older versions that were using this API to stop functioning which meant that our operator could not be installed. Through manual verification I have checked that the minimum cert-manager version that satisfies the 1.22.5 changes is cert-manager:v1.3.0 and I have added the findings in the installation guide.