Fixed wrongly storing mutated CRDs in kubernetes instead of the initial ones by using a deepcopy that is mutated and then sent to the gateway.
Due to using a single ApiDefinition struct for storing both the “input” and “output” CRD values, at some point if we want to work with values that are not going to be passed to the gateway/dashboard (as they would fail the api structure validator) we used to mutate the ApiDefinitions obtained from kubernetes directly. This behaviour led to the kubernetes API being unable to recognize whether a CRD was created or updated (as the stored CRD would be different from the one that was applied initially).