feat(helm): Add hostNetwork option (#537)
Add the option to turn host networking on.  With the host networking option on it is likely users will also want to control port selection so as not to collide with other pods running with hostNetwork mode enabled.  To accomodate port selection the ports have been parameterized in the template as well

Co-authored-by: Pranshu <104971506+singhpr@users.noreply.github.com>
Co-authored-by: Burak Sekili <buraksekili@gmail.com>